test(core) astubbs#317: split the 1456-line seam test into five classes on its own seams
The review's remaining two mechanical findings (#2, #11). MockConsumerCommitFailureSeamTest
had grown to 1456 lines and 27 flat test methods, and PMD CPD/jscpd flagged its repeated
setup and Awaitility blocks as this PR's largest new duplication.

Now: MockConsumerCommitFailureSeamTestBase owns the fixture (failing MockConsumer factory,
RecordingHandler, startPc overloads, latching feed, teardown, awaitAsserted and the CONTINUE
opening helpers), and five classes own one seam each - Decision (10), HandlerFreeExits (4),
PauseIntake (5), DeferralEscalation (5), Metrics (3). All 27 tests survive with assertions
byte-for-byte intact; the base keeps the documented decision NOT to extend
MockConsumerTestBase. ParallelConsumerOptionsCommitFailureTest's twin assertThrows blocks
collapse into one assertAsyncRejects helper. The deliberate
ConsumerManagerCommitRetryBudgetTest/ProducerManagerCommitBudgetTest mirror is untouched.

Verified: 27/27 across the five classes plus options/symptom suites, maven exit 0;
copyright, issue-ref and file-ref gates green.
